// smartAvc-new/src/app/shared/models/transporter.ts
export class Transporter {
         first_name: string;
         last_name: string;
         national_id: number;
         province: string;
         district: string;
         sector: string;
         phone_number_mtn: string= "078";
         phone_number_airtel: string="073";
         phone_number_tigo: string="072";
         gender: string;
         age: number;
         married: string;
         vehicle_type: string;
         plate_number: string;
}